Refactoring: 자잘한 경로 업데이트, cascade 옵션 변경
