+++ src/firebase/firebase.controller.spec.ts
@@ -1,7 +1,7 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { FirebaseController } from './firebase.controller';
+import { Test, TestingModule } from "@nestjs/testing";
+import { FirebaseController } from "./firebase.controller";
 
-describe('FirebaseController', () => {
+describe("FirebaseController", () => {
   let controller: FirebaseController;
 
   beforeEach(async () => {
@@ -12,7 +12,7 @@ describe('FirebaseController', () => {
     controller = module.get<FirebaseController>(FirebaseController);
   });
 
-  it('should be defined', () => {
+  it("should be defined", () => {
     expect(controller).toBeDefined();
   });
 });
